Sprint 5

See merge request 2021ss-sepm-pr-group/21ss-sepm-pr-inso-08!64

+++ backend/src/test/java/at/ac/tuwien/sepm/groupphase/backend/basetest/TestDataUser.java
@@ -16,7 +16,7 @@ public interface TestDataUser extends TestData{
     String ADMIN_EMAIL = "admin@email.com";
     String ADMIN_PASSWORD = "password";
     LocalDateTime ADMIN_LAST_LOGIN = LocalDateTime.now();
-    int ADMIN_POINTS = 0;
+    Long ADMIN_POINTS = 0L;
     ApplicationUser.UserStatus ADMIN_USER_STATUS = ApplicationUser.UserStatus.ACTIVE;
     ApplicationUser.UserRole ADMIN_ROLE = ApplicationUser.UserRole.ADMIN;
     Address ADMIN_ADDRESS = Address.builder().name("Max Mustermann").lineOne("Teststraße 1").city("Wien").postcode("1010").country("Österreich").build();
@@ -27,7 +27,7 @@ public interface TestDataUser extends TestData{
     String DEFAULT_EMAIL = "user@email.com";
     String DEFAULT_PASSWORD = "password";
     LocalDateTime DEFAULT_LAST_LOGIN = LocalDateTime.now();
-    int DEFAULT_POINTS = 0;
+    Long DEFAULT_POINTS = 0L;
     ApplicationUser.UserStatus DEFAULT_STATUS = ApplicationUser.UserStatus.ACTIVE;
     ApplicationUser.UserRole DEFAULT_ROLE = ApplicationUser.UserRole.CLIENT;
     Address DEFAULT_ADDRESS = Address.builder().name("Max Mustermann").lineOne("Teststraße 1").city("Wien").postcode("1010").country("Österreich").build();
